+++ .php-cs-fixer.dist.php
@@ -2,15 +2,27 @@
 
 $finder = PhpCsFixer\Finder::create()
     ->in(['src', 'tests'])
+    ->exclude('var/cache')
 ;
 
 $config = new PhpCsFixer\Config();
 return $config
-    // keep close to the Symfony standard but allow
-    // alignment of keys/values in array definitions
     ->setRules([
+        // keep close to the Symfony standard
         '@Symfony'               => true,
-        'binary_operator_spaces' => false,
+
+        // but force alignment of keys/values in array definitions
+        'binary_operator_spaces' => [
+            'operators' => [
+                '=>' => 'align_single_space_minimal_by_scope',
+                '='  => null,
+            ],
+        ],
+
+        // this would otherwise separate annotations
+        'phpdoc_separation'      => [
+            'skip_unlisted_annotations' => true,
+        ],
     ])
     ->setFinder($finder)
 ;