+++ tests/Filter/JsonExistsFilterTest.php
@@ -4,12 +4,14 @@
 
 declare(strict_types=1);
 
+namespace Vrok\SymfonyAddons\Tests\Filter;
+
 use ApiPlatform\Doctrine\Orm\Util\QueryNameGenerator;
 use ApiPlatform\Metadata\Get;
 use Doctrine\ORM\QueryBuilder;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Vrok\SymfonyAddons\Filter\JsonExistsFilter;
-use Vrok\SymfonyAddons\Tests\Fixtures\TestEntity;
+use Vrok\SymfonyAddons\Tests\Fixtures\Entity\TestEntity;
 
 /**
  * @group JsonExistsFilter
@@ -22,7 +24,7 @@ class JsonExistsFilterTest extends KernelTestCase
         $filter = new JsonExistsFilter($doctrine, null, ['jsonColumn' => null], null);
 
         $this->assertEquals([
-            'jsonColumn' => [
+            'jsonColumn'   => [
                 'property' => 'jsonColumn',
                 'type'     => 'string',
                 'required' => false,