+++ tests/Fixtures/app/AppKernel.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 use ApiPlatform\Symfony\Bundle\ApiPlatformBundle;
 use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
+use Doctrine\Bundle\FixturesBundle\DoctrineFixturesBundle;
 use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
 use Symfony\Bundle\FrameworkBundle\Kernel\MicroKernelTrait;
 use Symfony\Bundle\MonologBundle\MonologBundle;
+use Symfony\Bundle\SecurityBundle\SecurityBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpFoundation\Session\SessionFactory;
@@ -19,10 +21,10 @@ class AppKernel extends Kernel
     public function registerBundles(): array
     {
         $bundles = [
-            new ApiPlatformBundle(),
+            new FrameworkBundle(),
             new DoctrineBundle(),
             new MonologBundle(),
-            new FrameworkBundle(),
+            new ApiPlatformBundle(),
         ];
 
         return $bundles;
@@ -40,17 +42,17 @@ class AppKernel extends Kernel
         $loader->load(__DIR__.'/config/config.yaml');
 
         $c->prependExtensionConfig('framework', [
-            'property_access' => ['enabled' => true],
-            'secret' => 'symfony.vrok',
-            'validation' => ['enable_annotations' => true],
-            'serializer' => ['enable_annotations' => true],
-            'test' => null,
-            'session' => class_exists(SessionFactory::class) ? ['storage_factory_id' => 'session.storage.factory.mock_file'] : ['storage_id' => 'session.storage.mock_file'],
-            'profiler' => [
+            'property_access'      => ['enabled' => true],
+            'secret'               => 'symfony.vrok',
+            'validation'           => ['enable_annotations' => true],
+            'serializer'           => ['enable_annotations' => true],
+            'test'                 => null,
+            'session'              => class_exists(SessionFactory::class) ? ['storage_factory_id' => 'session.storage.factory.mock_file'] : ['storage_id' => 'session.storage.mock_file'],
+            'profiler'             => [
                 'enabled' => true,
                 'collect' => false,
             ],
-            'router' => ['utf8' => true],
+            'router'               => ['utf8' => true],
             'http_method_override' => false,
         ]);
     }